requirements.txt
#For development
fastapi
uvicorn
httpx
qrcode
psycopg2
enum34
redis==4.5.5
Pillow

python-jose[cryptography]
passlib[bcrypt]
python-multipart

sqlmodel
alembic

#For tests
pytest==7.2.0
pytest-cov==2.11.1
pytest-mock==3.5.1
mock==4.0.3
pytest-asyncio==0.20.1
pydantic-factories==1.13.0
pytest_httpx==0.21.2
